move make_cluster_overlay_img
