+++ napari_clusters_plotter/_plotter.py
@@ -8,7 +8,6 @@ import numpy as np
 import pandas as pd
 from matplotlib.figure import Figure
 from napari_tools_menu import register_dock_widget
-from PIL import ImageColor
 from qtpy import QtWidgets
 from qtpy.QtCore import Qt
 from qtpy.QtGui import QGuiApplication, QIcon
@@ -29,6 +28,7 @@ from ._plotter_utilities import (
     clustered_plot_parameters,
     estimate_number_bins,
     unclustered_plot_parameters,
+    make_cluster_overlay_img
 )
 from ._Qt_code import (
     ICON_ROOT,
@@ -425,42 +425,6 @@ class PlotterWidget(QMainWindow):
         self.plot_y_axis.setCurrentIndex(former_y_axis)
         self.plot_cluster_id.setCurrentIndex(former_cluster_id)
 
-    def make_cluster_overlay_img(
-        self,
-        cluster_id: str,
-        features: pd.DataFrame,
-        histogram_data: typing.Tuple,
-        feature_x: str,
-        feature_y: str,
-        colors: List[str],
-    ) -> np.array:
-        h, xedges, yedges = histogram_data
-
-        relevant_entries = features.loc[
-            features[cluster_id] != features[cluster_id].min(),
-            [cluster_id, feature_x, feature_y],
-        ]
-
-        cluster_overlay_rgba = np.zeros((*h.shape, 4), dtype=float)
-        output_max = np.zeros(h.shape, dtype=float)
-
-        for cluster, entries in relevant_entries.groupby(cluster_id):
-            h2, _, _ = np.histogram2d(
-                entries[feature_x], entries[feature_y], bins=[xedges, yedges]
-            )
-            mask = h2 > output_max
-            np.maximum(h2, output_max, out=output_max)
-            rgba = [
-                float(v) / 255
-                for v in list(
-                    ImageColor.getcolor(colors[int(cluster) % len(colors)], "RGB")
-                )
-            ]
-            rgba.append(0.9)
-            cluster_overlay_rgba[mask] = rgba
-
-        return cluster_overlay_rgba.swapaxes(0, 1)
-
     def run(
         self,
         features: pd.DataFrame,
@@ -564,7 +528,7 @@ class PlotterWidget(QMainWindow):
                     log_scale=self.log_scale.isChecked(),
                 )
 
-                rgb_img = self.make_cluster_overlay_img(
+                rgb_img = make_cluster_overlay_img(
                     cluster_id=plot_cluster_name,
                     features=features,
                     feature_x=self.plot_x_axis_name,
